Feature/exercise

+++ controllers/groups/delete.group.js
@@ -1,15 +1,21 @@
 const { AsyncCatch } = require("../../helpers/utils.helper");
 const { DefaultError, NotFound, Unauthorized } = require("../../helpers/errors.helper");
 const { Group } = require("../../models/Group.model");
+const { Course } = require("../../models/Course.model");
 const validator = require("../../helpers/validator.helper");
 const validatorSchema = require("../../validators/group.validator");
 
 module.exports = AsyncCatch(async (req, res, next) => {
-    const group = await Group.findById(req.params.id);
+    const params = validator(validatorSchema(["id"]), req.params);
+
+    const group = await Group.findById(params.id);
     if (!group) throw new NotFound("Not found.");
 
     const input = validator(validatorSchema(["name"]), req.body);
     if (input.name !== group.name) throw new Unauthorized("Name is not correct.");
+
+    const course = await Course.updateOne({ name: group.course }, { $pull: { groups: group._id } });
+    if (!course) throw new DefaultError("Can't connect to database.");
 
     const result = await Group.findOneAndDelete({ name: input.name });
     if (!result) throw new DefaultError("Can't connect to database.");
